tcpclose event_routes issue
